+++ R/ragnar-package.R
@@ -12,14 +12,13 @@ ragnartools.markitdown <- NULL
     ORT_DISABLE_TELEMETRY = "1"
   )
   S7::methods_register()
-  reticulate::py_require(c(
-    # Pin onnxruntime until this is resolved:
-    #  https://github.com/microsoft/markitdown/issues/1266
-    # New VC++ version requirement begins:
-    # https://github.com/Microsoft/onnxruntime/releases/tag/v1.21.0
-    if (is_windows()) "onnxruntime<=1.20.1",
-    "markitdown[all]"
-  ))
+  reticulate::py_require(
+    "markitdown[all]",
+    # Magika 0.6.3 (via MarkItDown) pins onnxruntime<=1.20.1 on Windows.
+    # That ONNX Runtime version has no Python 3.14 wheels.
+    # https://github.com/posit-dev/raghilda/issues/93
+    python_version = "<3.14"
+  )
 
   uv_set_override_latest_youtube_transcript_api()